pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "feedback-router"
version = "0.1.0"
description = "Intelligent feedback routing system for Gen X professionals"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "ItsJen Team", email = "engineering@itjen.ai"}
]
keywords = ["feedback", "routing", "ai", "automation", "gen-x"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Email",
    "Topic :: Office/Business",
    "Topic :: Software :: Libraries :: Application Frameworks",
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.9",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.26.0",
    "anthropic>=0.8.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.0.0",
    "python-json-logger>=2.0.0",
    "cryptography>=41.0.0",
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.0",
    "sentry-sdk>=1.39.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "pre-commit>=3.5.0",
]
google = [
    "google-auth>=2.25.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=1.12.0",
]
monitoring = [
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-exporter-otlp>=0.40b0",
    "datadog>=0.47.0",
]

[project.urls]
Homepage = "https://itjen.ai"
Repository = "https://github.com/itjen/feedback-router"
Documentation = "https://github.com/itjen/feedback-router/tree/main/docs"
Issues = "https://github.com/itjen/feedback-router/issues"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
py_version = 311
skip_gitignore = true
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
plugins = [
    "pydantic.mypy"
]

[tool.pydantic]
[tool.pydantic.mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "class .*\\(Protocol\\):",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",      # Pycodestyle errors
    "F",      # Pyflakes
    "I",      # isort
    "C90",    # mccabe
    "N",      # pep8-naming
    "D",      # pydocstyle
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "A",      # flake8-builtins
    "PIE",    # flake8-pie
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "TCH",    # flake8-type-checking
]
ignore = [
    "D100",   # Missing docstring in public module
    "D104",   # Missing docstring in public package
    "E501",   # Line too long
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["D"]

[tool.ruff.pydocstyle]
convention = "google"
